# openapi.yaml
openapi: 3.0.0
info:
  title: Learnify API
  version: 1.0.0
  description: |
    Learnify is a personalized learning platform API that provides endpoints for user authentication,
    course management, exercise submissions, social features, and AI-powered code reviews.

    ## Authentication
    All protected endpoints require a JWT Bearer token in the Authorization header:
    `Authorization: Bearer <token>`

    ## Features
    - User registration and authentication
    - Personalized learning paths
    - AI-powered code reviews
    - Social features (follow, activity feed)
    - Achievement system
    - Course recommendations

  contact:
    name: Learnify API Support
    email: support@learnify.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.learnify.example.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request body"
      required:
        - error
...
    OnboardingRequest:
      type: object
      required:
        - meta_category
        - domain
        - skill_level
      properties:
        meta_category:
          type: string
          example: "programming"
          description: High-level category of learning interest
        domain:
          type: string
          example: "web-development"
          description: Specific domain within the category
        skill_level:
          type: string
          enum: [beginner, intermediate, advanced]
          example: "beginner"
        variables:
          type: object
          additionalProperties:
            type: string
          example:
            preferred_language: "javascript"
            learning_goal: "build-web-apps"
    Course:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "456e7890-e89b-12d3-a456-426614174001"
        user_id:
          type: string
          format: uuid
        archetype_id:
          type: string
          format: uuid
        title:
          type: string
          example: "Modern Web Development with React"
        description:
          type: string
          example: "Learn to build modern web applications using React, Node.js, and PostgreSQL"
        meta_category:
          type: string
          example: "programming"
        status:
          type: string
          enum: [active, completed, archived]
          example: "active"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Module:
      type: object
      properties:
        id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
        module_number:
          type: integer
          example: 1
        title:
          type: string
          example: "Introduction to React Components"
        description:
          type: string
          example: "Learn the fundamentals of React components and JSX"
        status:
          type: string
          enum: [locked, unlocked, completed]
          example: "unlocked"
        unlocked_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    Exercise:
      type: object
      properties:
        id:
          type: string
          format: uuid
        module_id:
          type: string
          format: uuid
        exercise_number:
          type: integer
          example: 1
        title:
          type: string
          example: "Build a Counter Component"
        description:
          type: string
          example: "Create a React component that increments and decrements a counter"
        language:
          type: string
          example: "javascript"
        starter_code:
          type: string
          example: "import React from 'react';\n\nexport default function Counter() {\n  // Your code here\n}"
        difficulty:
          type: string
          enum: [easy, medium, hard]
          example: "easy"
        points:
          type: integer
          example: 100
        hints:
          type: array
          items:
            type: string
          example:
            - "Use the useState hook"
            - "Create two button elements for increment and decrement"
        created_at:
          type: string
          format: date-time
    SubmitExerciseRequest:
      type: object
      required:
        - code
        - language
      properties:
        code:
          type: string
          example: "import React, { useState } from 'react';\n\nexport default function Counter() {\n  const [count, setCount] = useState(0);\n  return (\n    <div>\n      <p>Count: {count}</p>\n      <button onClick={() => setCount(count + 1)}>+</button>\n      <button onClick={() => setCount(count - 1)}>-</button>\n    </div>\n  );\n}"
        language:
          type: string
          example: "javascript"
    ModuleCompletion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        module_id:
          type: string
          format: uuid
        exercise_id:
          type: string
          format: uuid
        submitted_code:
          type: string
        language:
          type: string
        passed:
          type: boolean
          example: true
        score:
          type: integer
          example: 95
        attempts:
          type: integer
          example: 2
        hints_used:
          type: integer
          example: 1
        time_spent_minutes:
          type: integer
          example: 15
        submitted_at:
          type: string
          format: date-time
    ArchitectureReview:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        module_id:
          type: string
          format: uuid
        submission_id:
          type: string
          format: uuid
        overall_score:
          type: integer
          minimum: 0
          maximum: 100
          example: 88
        code_sense_score:
          type: integer
          example: 90
        efficiency_score:
          type: integer
          example: 85
        edge_cases_score:
          type: integer
          example: 92
        taste_score:
          type: integer
          example: 86
        feedback:
          type: object
          properties:
            strengths:
              type: array
              items:
                type: string
              example:
                - "Clean component structure"
                - "Proper use of React hooks"
            improvements:
              type: array
              items:
                type: string
              example:
                - "Consider adding prop validation"
                - "Extract button logic to separate component"
            suggestions:
              type: string
              example: "Great job! Consider learning about useReducer for more complex state management."
        reviewed_at:
          type: string
          format: date-time
    UserProgress:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
        current_module_id:
          type: string
          format: uuid
        progress_percentage:
          type: integer
          minimum: 0
          maximum: 100
          example: 45
        time_spent_minutes:
          type: integer
          example: 320
        last_activity:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
...
paths:
...
  /api/onboarding/complete:
    post:
      tags:
        - Users
      summary: Complete user onboarding
      description: Sets user's learning preferences and generates personalized course
      operationId: completeOnboarding
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingRequest'
      responses:
        '200':
          description: Onboarding completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "onboarding completed successfully"
        '400':
          description: Invalid request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "meta_category, domain, and skill_level are required"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/courses:
    get:
      tags:
        - Courses
      summary: Get user's courses
      description: Retrieves all courses for the authenticated user
      operationId: getUserCourses
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Courses retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Course'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
  /api/courses/{id}:
    get:
      tags:
        - Courses
      summary: Get course details
      description: Retrieves detailed information about a specific course including modules
      operationId: getCourseDetails
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Course UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Course details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      course:
                        $ref: '#/components/schemas/Course'
                      modules:
                        type: array
                        items:
                          $ref: '#/components/schemas/Module'
        '400':
          description: Invalid course ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
  /api/courses/{id}/progress:
    get:
      tags:
        - Courses
      summary: Get course progress
      description: Retrieves the user's progress for a specific course
      operationId: getCourseProgress
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Course UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Progress retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/UserProgress'
        '400':
          description: Invalid course ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '404':
          description: Progress not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
  /api/exercises/{id}:
    get:
      tags:
        - Exercises
      summary: Get exercise details
      description: Retrieves details for a specific exercise
      operationId: getExercise
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Exercise UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Exercise retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Exercise'
        '400':
          description: Invalid exercise ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '404':
          description: Exercise not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
  /api/exercises/{id}/submit:
    post:
      tags:
        - Exercises
      summary: Submit exercise solution
      description: Submits code for an exercise and receives automated feedback
      operationId: submitExercise
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Exercise UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitExerciseRequest'
      responses:
        '200':
          description: Exercise submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ModuleCompletion'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
              examples:
                missingCode:
                  value:
                    error: "Bad Request"
                    message: "Code is required"
                missingLanguage:
                  value:
                    error: "Bad Request"
                    message: "Language is required"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
  /api/submissions/{id}/review:
    post:
      tags:
        - Exercises
      summary: Request AI code review
      description: Requests an AI-powered architecture review for a submission
      operationId: requestReview
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Submission UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Review completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ArchitectureReview'
        '400':
          description: Invalid submission ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
...